requirement.txt
flask
Flask-SQLAlchemy
flask-login
python-dotenv    